Fixing an issue with the validateEntries function being undefined
